+++ src/crosschain-data/adapters/layerzero/LayerzeroImplementation.sol
@@ -100,7 +100,7 @@ contract LayerzeroImplementation is IAmbImplementation, ILayerZeroUserApplicatio
         /// @dev  reset old mappings
         uint64 oldSuperChainId = superChainId[ambChainId_];
 
-        uint64 oldAmbChainId = ambChainId[superChainId_];
+        uint16 oldAmbChainId = ambChainId[superChainId_];
 
         if (oldSuperChainId > 0) {
             ambChainId[oldSuperChainId] = 0;
